requirements.txt
pytest
fastapi
email-validator
python-dotenv
requests
pydantic

psycopg2-binary
sqlalchemy
sqlmodel

fastapi-pagination
faker
